+++ src/main/java/string_calculator/StringAddCalculator.java
@@ -1,4 +1,4 @@
-package StringCalculator;
+package string_calculator;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
